[#613] Moved testhelper classes to appropiate package
