+++ app/src/androidTest/java/org/akvo/flow/activity/FreeTextFormTest.java
@@ -30,8 +30,8 @@ import android.support.test.runner.AndroidJUnit4;
 import org.akvo.flow.R;
 import org.akvo.flow.data.database.SurveyDbAdapter;
 import org.akvo.flow.domain.Survey;
-import org.akvo.flow.testhelper.SurveyInstaller;
-import org.akvo.flow.testhelper.SurveyRequisite;
+import org.akvo.flow.activity.testhelper.SurveyInstaller;
+import org.akvo.flow.activity.testhelper.SurveyRequisite;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
