+++ app/src/androidTest/java/org/akvo/flow/activity/OptionsSurveyTest.java
@@ -29,8 +29,8 @@ import android.support.test.runner.AndroidJUnit4;
 import org.akvo.flow.R;
 import org.akvo.flow.data.database.SurveyDbAdapter;
 import org.akvo.flow.domain.Survey;
-import org.akvo.flow.testhelper.SurveyInstaller;
-import org.akvo.flow.testhelper.SurveyRequisite;
+import org.akvo.flow.activity.testhelper.SurveyInstaller;
+import org.akvo.flow.activity.testhelper.SurveyRequisite;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
